pyproject.toml
[project]
name = "django-ptb"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Nafnix", email = "nafnix@proton.me"},
]
dependencies = [
    "django>=5.0.3",
    "pydantic-settings>=2.2.1",
    "django-debug-toolbar>=4.3.0",
    "python-telegram-bot>=21.0.1",
    "uvicorn>=0.29.0",
    "gunicorn>=21.2.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false


[tool.pdm.dev-dependencies]
dev = [
    "pre-commit>=3.6.2",
    "ruff>=0.3.3",
    "devtools>=0.12.2",
]
[tool.ruff]
line-length = 99
fix = true

[tool.ruff.format]
quote-style = "single"
skip-magic-trailing-comma = false
docstring-code-format = true

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "E", # pycodestyle
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "A", # flake8-builtins,
    "DJ", # flake8-django
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
]
ignore = ["E111", "E114", "E117"]

# https://docs.astral.sh/ruff/settings/#lintflake8-errmsg
# Maximum string length for string literals in exception messages.
[tool.ruff.lint.flake8-errmsg]
max-string-length = 20

# https://docs.astral.sh/ruff/settings/#lintisort
[tool.ruff.lint.isort]
case-sensitive = true
lines-after-imports = 2

# https://docs.astral.sh/ruff/settings/#lint_pycodestyle_max-doc-length
[tool.ruff.lint.pycodestyle]
max-doc-length = 72
